Finished creating .sh script and calling at the beginning of main to serve our static data to the public folders

+++ prep_public.sh
@@ -1,1 +1,5 @@
-rm -r ./public/*
+if [ -d "./public/*" ]; then
+    rm -r ./public/*
+    echo "Removed all files and directories in public/"
+fi
+cp -a ./static/* ./public/
